+++ scraperweb/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["__version__"]
 
-__version__ = "1.32.0"
+__version__ = "1.32.1"